login modal, JS overhaul, readme, misc fixes
